# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Rocket Trading API
  version: 1.1.0
  description: |
    Contract for the existing Spring Boot demo API. JSON numbers represent
    BigDecimal values; timestamps are UTC instants. Authentication is currently
    email-only demo sign-in. See README.md for implementation gaps and usage.
servers:
  - url: http://localhost:8081/api/v1
    description: Local backend
security:
  - bearerAuth: []
tags:
  - name: Auth
  - name: Profile
  - name: Instruments
  - name: Quotes
  - name: Portfolio
  - name: Orders
  - name: Reporting
paths:
  /auth/register:
    post:
      tags: [Auth]
      operationId: registerClient
      summary: Register a client and a USD trading account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterClientRequest'
            example: {name: Demo Trader, email: trader@example.com, initialCash: 10000}
      responses:
        '201':
          description: Client registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/sign-in:
    post:
      tags: [Auth]
      operationId: signIn
      summary: Create a session using a registered email
      description: Demo authentication; no password is accepted or checked.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInRequest'
            example: {email: trader@example.com}
      responses:
        '200':
          description: Session and bearer token created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/sign-out:
    post:
      tags: [Auth]
      operationId: signOut
      summary: Expire the authenticated session
      responses:
        '204':
          description: Signed out; no response body
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /quotes/{symbol}:
    get:
      tags: [Quotes]
      operationId: getQuote
      summary: Retrieve a current quote from Fauxnance
      parameters:
        - name: symbol
          in: path
          required: true
          schema: {type: string}
          example: AAPL
        - name: market
          in: query
          description: stock or crypto, case-insensitive
          schema: {type: string, default: stock}
      responses:
        '200':
          description: Current quote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /portfolio/summary:
    get:
      tags: [Portfolio]
      operationId: getPortfolioSummary
      summary: Get the authenticated client's cash and positions
      responses:
        '200':
          description: Portfolio summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /me:
    get:
      tags: [Profile]
      operationId: getProfile
      summary: Get the signed-in fixture client profile
      responses:
        '200':
          description: Client profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /instruments:
    get:
      tags: [Instruments]
      operationId: listSupportedInstruments
      summary: List the supported tradable catalogue
      responses:
        '200':
          description: Supported instruments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstrumentListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /orders:
    post:
      tags: [Orders]
      operationId: submitOrder
      summary: Accept an order for asynchronous execution
      description: |
        Valid orders commit as ACCEPTED before a background worker executes them.
        Poll GET /orders or GET /orders/{orderId} for FILLED or REJECTED.
        Business rejections return 409 ORDER_REJECTED and remain in order history.
        Repeating a client-scoped Idempotency-Key with equivalent normalized details
        returns the existing order (202, or 409 if rejected); changed details return
        409 IDEMPOTENCY_CONFLICT. Reuse the key after network errors.
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, minLength: 1, maxLength: 100, pattern: '^[A-Za-z0-9._:-]+$'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitOrderRequest'
            example: {symbol: AAPL, side: BUY, quantity: 1, market: stock, orderType: MARKET}
      responses:
        '202':
          description: Accepted order, or the current state of an idempotent retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Orders]
      operationId: listOrders
      summary: List the authenticated client's orders
      description: Returns an array without pagination.
      responses:
        '200':
          description: Orders, possibly empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /orders/{orderId}:
    get:
      tags: [Orders]
      operationId: getOrder
      summary: Get an order belonging to the authenticated client
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /orders/{orderId}/timeline:
    get:
      tags: [Orders]
      operationId: getOrderTimeline
      summary: Get the order's audit timeline for the authenticated client
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Ordered audit events for submission, pricing, settlement and fills
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderTimelineEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /fills/{orderId}:
    get:
      tags: [Orders]
      operationId: listFills
      summary: List fills for an order
      description: |
        Only the authenticated owner can read fills. Unknown orders and orders
        belonging to another client both return 404.
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Fills, possibly empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FillListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /reporting/overview:
    get:
      tags: [Reporting]
      operationId: getReportingOverview
      summary: Get a reporting summary for a time window
      parameters:
        - name: from
          in: query
          required: false
          schema: {type: string, format: date-time}
        - name: to
          in: query
          required: false
          schema: {type: string, format: date-time}
      responses:
        '200':
          description: Aggregate reporting snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportingOverviewEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /reporting/instruments:
    get:
      tags: [Reporting]
      operationId: listInstrumentActivity
      summary: Break out trading activity by instrument over a time window
      parameters:
        - name: from
          in: query
          required: false
          schema: {type: string, format: date-time}
        - name: to
          in: query
          required: false
          schema: {type: string, format: date-time}
      responses:
        '200':
          description: Per-instrument activity summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstrumentActivityListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /reporting/segments:
    get:
      tags: [Reporting]
      operationId: listClientSegmentActivity
      summary: Break out trading activity by client risk profile over a time window
      parameters:
        - name: from
          in: query
          required: false
          schema: {type: string, format: date-time}
        - name: to
          in: query
          required: false
          schema: {type: string, format: date-time}
      responses:
        '200':
          description: Per-segment activity summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientSegmentActivityListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token from sign-in; the corresponding server-side session must remain valid.
  parameters:
    OrderId:
      name: orderId
      in: path
      required: true
      schema: {type: integer, format: int64}
  responses:
    BadRequest:
      description: VALIDATION_FAILED; invalid fields, JSON, or parameter types
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiErrorResponse'}
    Unauthorized:
      description: UNAUTHENTICATED or INVALID_CREDENTIALS
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiErrorResponse'}
    NotFound:
      description: ACCOUNT_NOT_FOUND, ORDER_NOT_FOUND, or INSTRUMENT_NOT_FOUND
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiErrorResponse'}
    Conflict:
      description: CLIENT_ALREADY_EXISTS, ORDER_REJECTED, IDEMPOTENCY_CONFLICT, QUOTE_STALE, QUOTE_INVALID, or CONFLICT
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiErrorResponse'}
    Unavailable:
      description: PRICE_UNAVAILABLE or QUOTE_PROVIDER_NOT_CONFIGURED
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiErrorResponse'}
    InternalError:
      description: INTERNAL_ERROR; unexpected server failure
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiErrorResponse'}
  schemas:
    RegisterClientRequest:
      type: object
      required: [name, email]
      properties:
        name: {type: string, minLength: 1, pattern: '\S'}
        email: {type: string, format: email, minLength: 1, pattern: '\S'}
        dateOfBirth: {type: string, format: date, nullable: true, default: '1990-01-01'}
        riskProfile:
          type: string
          nullable: true
          default: Balanced
          description: Omitted, null, or blank uses Balanced; otherwise trimmed.
        initialCash: {type: number, minimum: 0.01, nullable: true, default: 10000}
    SignInRequest:
      type: object
      required: [email]
      properties:
        email: {type: string, format: email, minLength: 1, pattern: '\S'}
    SubmitOrderRequest:
      type: object
      required: [symbol, side, quantity]
      description: Invalid or missing fields return 400. LIMIT requires a positive limitPrice; MARKET must omit it or send null.
      properties:
        symbol: {type: string, minLength: 1, maxLength: 40, pattern: '\S', example: AAPL}
        side:
          type: string
          minLength: 1
          pattern: '\S'
          description: BUY or SELL, trimmed and case-insensitive.
          example: BUY
        quantity:
          type: number
          minimum: 0.000001
          maximum: 999999999999.999999
          multipleOf: 0.000001
          description: At most twelve integer digits and six fractional digits; excess precision is rejected.
        market:
          type: string
          nullable: true
          default: stock
          description: stock or crypto, trimmed and case-insensitive; null or blank uses stock.
        orderType:
          type: string
          nullable: true
          default: MARKET
          description: MARKET or LIMIT, trimmed and case-insensitive; null or blank uses MARKET.
        limitPrice:
          type: number
          nullable: true
          minimum: 0.0001
          maximum: 999999999999.9999
          multipleOf: 0.0001
          description: Required for LIMIT; null or omitted for MARKET. Checked at acceptance and execution.
    ClientRegistrationResponse:
      type: object
      required: [clientId, name, email, registeredAt]
      properties:
        clientId: {type: integer, format: int64}
        name: {type: string}
        email: {type: string}
        registeredAt: {type: string, format: date-time}
    SessionResponse:
      type: object
      required: [clientId, sessionId, expiresAt, accessToken, tokenType]
      properties:
        clientId: {type: integer, format: int64}
        sessionId: {type: integer, format: int64}
        expiresAt: {type: string, format: date-time}
        accessToken: {type: string}
        tokenType: {type: string, enum: [Bearer]}
    QuoteResponse:
      type: object
      required: [symbol, bid, ask, price, bidVolume, askVolume, capturedAt, market]
      properties:
        symbol: {type: string}
        bid: {type: number}
        ask: {type: number}
        price: {type: number}
        bidVolume: {type: integer, format: int64}
        askVolume: {type: integer, format: int64}
        capturedAt: {type: string, format: date-time}
        market: {type: string}
    OrderResponse:
      type: object
      required: [orderId, symbol, side, orderType, quantity, status, rejectionReason, submittedAt]
      properties:
        orderId: {type: integer, format: int64}
        symbol: {type: string}
        side: {type: string}
        orderType: {type: string}
        quantity: {type: number}
        status: {type: string, example: FILLED}
        rejectionReason: {type: string, nullable: true}
        submittedAt: {type: string, format: date-time}
    FillResponse:
      type: object
      required: [fillId, orderId, executedQuantity, executedPrice, executedAt]
      properties:
        fillId: {type: integer, format: int64}
        orderId: {type: integer, format: int64}
        executedQuantity: {type: number}
        executedPrice: {type: number}
        executedAt: {type: string, format: date-time}
    PositionResponse:
      type: object
      required: [symbol, quantity]
      properties:
        symbol: {type: string}
        quantity: {type: number}
    PortfolioSummaryResponse:
      type: object
      required: [clientId, accountId, cashBalance, currency, positions]
      properties:
        clientId: {type: integer, format: int64}
        accountId: {type: integer, format: int64}
        cashBalance: {type: number}
        currency: {type: string, example: USD}
        positions:
          type: array
          items: {$ref: '#/components/schemas/PositionResponse'}
    ClientProfileResponse:
      type: object
      required: [clientId, name, email]
      properties:
        clientId: {type: integer, format: int64}
        name: {type: string}
        email: {type: string, format: email}
    SupportedInstrumentResponse:
      type: object
      required: [instrumentId, symbol, name, assetClass, baseCurrency, tradable, market]
      properties:
        instrumentId: {type: integer, format: int64}
        symbol: {type: string}
        name: {type: string}
        assetClass: {type: string}
        baseCurrency: {type: string}
        tradable: {type: boolean}
        market: {type: string, enum: [stock, crypto]}
    OrderTimelineEventResponse:
      type: object
      required: [auditId, entityName, entityId, actionType, recordedAt, stateAfter]
      properties:
        auditId: {type: integer, format: int64}
        entityName: {type: string}
        entityId: {type: integer, format: int64}
        actionType: {type: string}
        recordedAt: {type: string, format: date-time}
        stateBefore: {type: object, nullable: true, additionalProperties: true}
        stateAfter: {type: object, additionalProperties: true}
    OrderTimelineResponse:
      type: object
      required: [orderId, events]
      properties:
        orderId: {type: integer, format: int64}
        events:
          type: array
          items: {$ref: '#/components/schemas/OrderTimelineEventResponse'}
    ReportingOverviewResponse:
      type: object
      required: [from, to, totalOrders, acceptedOrders, filledOrders, rejectedOrders, totalFills, totalNotional, activeClients]
      properties:
        from: {type: string, format: date-time}
        to: {type: string, format: date-time}
        totalOrders: {type: integer, format: int64}
        acceptedOrders: {type: integer, format: int64}
        filledOrders: {type: integer, format: int64}
        rejectedOrders: {type: integer, format: int64}
        totalFills: {type: integer, format: int64}
        totalNotional: {type: number}
        activeClients: {type: integer, format: int64}
    InstrumentActivityResponse:
      type: object
      required: [symbol, assetClass, orderCount, fillCount, buyQuantity, sellQuantity]
      properties:
        symbol: {type: string}
        assetClass: {type: string}
        orderCount: {type: integer, format: int64}
        fillCount: {type: integer, format: int64}
        buyQuantity: {type: number}
        sellQuantity: {type: number}
        lastSubmittedAt: {type: string, format: date-time, nullable: true}
    ClientSegmentActivityResponse:
      type: object
      required: [riskProfile, clientCount, orderCount, totalNotional]
      properties:
        riskProfile: {type: string}
        clientCount: {type: integer, format: int64}
        orderCount: {type: integer, format: int64}
        totalNotional: {type: number}
    ResponseMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId: {type: string, format: uuid}
        timestamp: {type: string, format: date-time}
    ApiErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details, requestId, timestamp]
          properties:
            code: {type: string, example: VALIDATION_FAILED}
            message: {type: string}
            details:
              type: array
              items:
                type: object
                required: [field, issue]
                properties:
                  field: {type: string}
                  issue: {type: string}
            requestId: {type: string, format: uuid}
            timestamp: {type: string, format: date-time}
    RegistrationEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {$ref: '#/components/schemas/ClientRegistrationResponse'}
        meta: {$ref: '#/components/schemas/ResponseMeta'}
    SessionEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {$ref: '#/components/schemas/SessionResponse'}
        meta: {$ref: '#/components/schemas/ResponseMeta'}
    QuoteEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {$ref: '#/components/schemas/QuoteResponse'}
        meta: {$ref: '#/components/schemas/ResponseMeta'}
    PortfolioEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {$ref: '#/components/schemas/PortfolioSummaryResponse'}
        meta: {$ref: '#/components/schemas/ResponseMeta'}
    ProfileEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {$ref: '#/components/schemas/ClientProfileResponse'}
        meta: {$ref: '#/components/schemas/ResponseMeta'}
    InstrumentListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: {$ref: '#/components/schemas/SupportedInstrumentResponse'}
        meta: {$ref: '#/components/schemas/ResponseMeta'}
    OrderEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {$ref: '#/components/schemas/OrderResponse'}
        meta: {$ref: '#/components/schemas/ResponseMeta'}
    OrderListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: {$ref: '#/components/schemas/OrderResponse'}
        meta: {$ref: '#/components/schemas/ResponseMeta'}
    FillListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: {$ref: '#/components/schemas/FillResponse'}
        meta: {$ref: '#/components/schemas/ResponseMeta'}
    OrderTimelineEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {$ref: '#/components/schemas/OrderTimelineResponse'}
        meta: {$ref: '#/components/schemas/ResponseMeta'}
    ReportingOverviewEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {$ref: '#/components/schemas/ReportingOverviewResponse'}
        meta: {$ref: '#/components/schemas/ResponseMeta'}
    InstrumentActivityListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: {$ref: '#/components/schemas/InstrumentActivityResponse'}
        meta: {$ref: '#/components/schemas/ResponseMeta'}
    ClientSegmentActivityListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: {$ref: '#/components/schemas/ClientSegmentActivityResponse'}
        meta: {$ref: '#/components/schemas/ResponseMeta'}